backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
idna==3.11
joblib==1.5.2
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-openid==2.2.5
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.6.0
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
